fixed and cleared all controllers

+++ src/main/java/com/vvelev/learnify/services/CourseService.java
@@ -4,7 +4,9 @@ import com.vvelev.learnify.dtos.course.CourseDto;
 import com.vvelev.learnify.dtos.course.CreateCourseDto;
 import com.vvelev.learnify.dtos.course.UpdateCourseDto;
 import com.vvelev.learnify.entities.Course;
+import com.vvelev.learnify.entities.EnrollmentId;
 import com.vvelev.learnify.entities.User;
+import com.vvelev.learnify.exceptions.AccessDeniedException;
 import com.vvelev.learnify.exceptions.CourseNotFoundException;
 import com.vvelev.learnify.exceptions.UserNotFoundException;
 import com.vvelev.learnify.mappers.CourseMapper;
@@ -16,6 +18,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @AllArgsConstructor
 @Service
@@ -23,6 +26,22 @@ public class CourseService {
     private final CourseRepository courseRepository;
     private final UserRepository userRepository;
     private final CourseMapper courseMapper;
+
+    public CourseDto createCourse(CreateCourseDto request) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Long id = (Long) authentication.getPrincipal();
+
+        User user = userRepository.findById(id).orElse(null);
+        if (user == null) {
+            throw new UserNotFoundException();
+        }
+
+        Course course = courseMapper.toEntity(request);
+        course.setCreatedBy(user);
+        courseRepository.save(course);
+
+        return courseMapper.toDto(course);
+    }
 
     public List<CourseDto> getAllCourses() {
         return courseRepository.findAll()
@@ -41,6 +60,13 @@ public class CourseService {
     }
 
     public List<CourseDto> getCoursesCreated(Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Long userId = (Long) authentication.getPrincipal();
+
+        if (!userId.equals(id)) {
+            throw new AccessDeniedException();
+        }
+
         User user = userRepository.findById(id).orElse(null);
         if (user == null) {
             throw new UserNotFoundException();
@@ -52,26 +78,14 @@ public class CourseService {
                 .toList();
     }
 
-    public CourseDto createCourse(CreateCourseDto request) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        Long id = (Long) authentication.getPrincipal();
-
-        User user = userRepository.findById(id).orElse(null);
-        if (user == null) {
-            throw new UserNotFoundException();
-        }
-
-        Course course = courseMapper.toEntity(request);
-        course.setCreatedBy(user);
-        courseRepository.save(course);
-
-        return courseMapper.toDto(course);
-    }
-
     public CourseDto updateCourse(Long id, UpdateCourseDto request) {
         Course course = courseRepository.findById(id).orElse(null);
         if (course == null) {
             throw new CourseNotFoundException();
+        }
+
+        if (!Objects.equals(course.getCreatedBy().getId(), id)) {
+            throw new AccessDeniedException();
         }
 
         courseMapper.update(request, course);
@@ -84,6 +98,10 @@ public class CourseService {
         Course course = courseRepository.findById(id).orElse(null);
         if (course == null) {
             throw new CourseNotFoundException();
+        }
+
+        if (!Objects.equals(course.getCreatedBy().getId(), id)) {
+            throw new AccessDeniedException();
         }
 
         courseRepository.delete(course);
